requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
qdrant-client>=1.8.0
feedparser>=6.0.11
google-generativeai>=0.4.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
beautifulsoup4>=4.12.3
